Split out the scheduler logic and define the APIs.
